- little changes to try to stabilize newgui (no luck yet)

+++ branches/newgui/mplayer/osdep/timer-gekko.c
@@ -23,6 +23,7 @@
 #include <unistd.h>
 
 #include <ogc/lwp_watchdog.h>
+#include <sys/time.h>
 
 const char *timer_name = "gekko";
 
@@ -30,15 +31,17 @@ int usec_sleep(int usec_delay) {
 	return usleep(usec_delay);
 }
 
+// Returns current time in microseconds
 u64 GetTimer(void) {
 	return ticks_to_microsecs(gettime());
 }
 
+// Returns current time in milliseconds
 u64 GetTimerMS(void) {
 	return ticks_to_millisecs(gettime());
 }
 
-//static u64 relative = 0;
+
 static u64 RelativeTime=0;
 
 double GetRelativeTime(void){
@@ -47,7 +50,7 @@ double GetRelativeTime(void){
 
   r=t-RelativeTime;
   RelativeTime=t;
-  return (float)r * 0.000001F;
+  return (double)r * 0.000001F;
 }
 /*
 float GetRelativeTime1(void) {
@@ -64,11 +67,13 @@ float GetRelativeTime1(void) {
 */
 static u64 timeinit;
 void InitTimer(void) {
+/*
 	timeinit = gettime();	
 	settime(timeinit); //testing
 	gettime();	
 	usleep(50);
 	gettime();	
+*/	
 	GetRelativeTime();
 	//relative = gettime();
 }
